reorganise package test
